backend/requirements_dev.txt
flake8==4.0.1
tox==3.25.0
pytest==7.1.2
pytest-cov==2.12.1
mypy==0.942
django-stubs==1.10.1
djangorestframework-stubs==1.5.0
drf-yasg-stubs==0.1.4
isort==5.10.1